+++ src/mpservice/server_process.py
@@ -1,1 +1,2 @@
+# This module is deprecated. Please import ``Manager`` from ``mpservice.multiprocessing``.
 from .multiprocessing import Manager