+++ src/mpservice/socket.py
@@ -16,7 +16,8 @@ from typing import Any, Awaitable, Callable, Optional
 
 from ._queues import SingleLane
 from .multiprocessing import RemoteException
-from .threading import MAX_THREADS, ThreadPoolExecutor
+from .threading import MAX_THREADS
+from .concurrent_futures import ThreadPoolExecutor
 from .util import is_async, get_docker_host_ip