+++ tests/test_multiprocessing_logger.py
@@ -20,7 +20,7 @@ def worker2():
 
 def test_logging():
     task1 = Process(target=worker1)
-    task2 = mp.Process(target=worker2)  # fork
+    task2 = mp.get_context('spawn').Process(target=worker2)  # fork
     task1.start()
     task2.start()
     task1.join()
@@ -30,3 +30,13 @@ def test_logging():
 if __name__ == '__main__':
     logging.basicConfig(level=logging.DEBUG)
     test_logging()
+
+# Printout is like this:
+#
+# worker2 warning
+# WARNING:worker1:worker1 warning
+# INFO:worker1:worker1 info
+# DEBUG:worker1:worker1 debug
+#
+# That is, in the standard spawned process, logging is not configured, hence
+# the printout is governed by the default in terms of both format and level.